TYPES: Add dashboard type
== Context

A while ago, when opex was thinking about all the types of
notifications, we thought having TYPE=dashboard would be useful.

The goal of TYPE=dashboard is to create a notification, without sending
an email, and that would purely be visible in the notifications tab.

== Why now?

SDDTLS (spaces-default-domain-tls) was recently launched. When you
create a private space app, we automatically provision a letsencrypt
TLS certificate. It would be great if we can create a notification as soon
as the cert is created, and uploaded to the app successfully -- without
sending an email.